Updated to prepare for RESP3 support
Added dummy RESP3 handle to command.go and rewrote command.go
and support code to simplify implementation of RESP3. Updated
tests to pass with changed data structures

+++ aggregatebuilder.go
@@ -53,7 +53,7 @@ func (a *AggregateBuilder) Verbatim() *AggregateBuilder {
 }
 
 // Limit sets the result limit
-func (a *AggregateBuilder) Limit(offset, num int) *AggregateBuilder {
+func (a *AggregateBuilder) Limit(offset, num int64) *AggregateBuilder {
 	a.opts.Steps = append(a.opts.Steps, &Limit{Offset: offset, Num: num})
 	return a
 }
